Ignore other values while waiting for mouse ACK
Some mice seem to send other values before we get the ACK. This was causing auxstate to be reset and the code was back to auxstate=0, waiting for 0xAA which will never happen. 

This change is to ignore other values and stay in auxstate=2 indefinitely until we get the ACK, which eventually comes. 

This makes the Logitech M-S69 mouse work (and possibly others).